refactoring.

+++ steph/src2/weight.h
@@ -1,7 +1,7 @@
 #ifndef ST_HEURISTIC_WEIGHT_H
 #define ST_HEURISTIC_WEIGHT_H
 
-typedef unsigned int weight_t;
+typedef unsigned long weight_t;
 
 extern int weight_compar(const void *p1, const void *p2);
